refactor(agents): own handle minting below store
