+++ AGENTS.md
@@ -103,7 +103,7 @@ This indexes what lives where; runtime shape and the single-binary rationale liv
 
 **Subsystems — `crates/rimz/src/`**, each carrying its own `AGENTS.md` contract:
 - `cli/` — command parsing, one `run(...)` per subcommand, shared `cli/render/` output, and the subagent fleet digest.
-- `agents/` — the provider-neutral `AgentDefinition` catalog, caller-aligned capability contracts and services, `state.rs` rollup, private `adapters/` implementations for every built-in and process plugin, and shared spend/pricing/account machinery including upstream pricing projection.
+- `agents/` — the provider-neutral `AgentDefinition` catalog, caller-aligned capability contracts and services, `state.rs` rollup, agent handle minting and validation (`petname`), private `adapters/` implementations for every built-in and process plugin, and shared spend/pricing/account machinery including upstream pricing projection.
 - `room/` — private managed-room context, birth/reset/teardown lifecycle, sidebar/presence options, and health gating.
 - `harness/` — layout IR, teams, address grammar, launch argv, supervised runs and their wake socket, loop scheduling over clock/signal/watch triggers, signal names and in-process firing, resume planning, and rebirth recovery inspection/materialization.
 - `message/` — message delivery: park-vs-live dispatch, live-pane send, reply waits, scheduled wakeups.
