+++ crates/rimz/src/agents/AGENTS.md
@@ -7,6 +7,7 @@ Topic detail lives in the internals leaves the root map describes — [adapter.m
 ## Layout
 
 - Shared, kind-agnostic code sits at the top level — [`definition.rs`](./definition.rs) owns the `AgentDefinition` catalog value and immutable `AgentSpec`, [`capabilities.rs`](./capabilities.rs) owns caller-aligned workflow contracts and the `AgentIntegration` bundle every adapter implements, and neutral services own lifecycle, session, context, runtime-control, account, transcript, and spending policy. Per-file detail lives in the `//!` headers.
+- [`petname.rs`](./petname.rs) owns the handle grammar: the adjective-noun minter with its deterministic fallback for pre-petname sessions, `valid_agent_name`, the reserved command words, and the pseudo-handles message envelopes give non-agent senders. It sits here because a handle is invalid when it reads as a catalog kind, which makes validity a question for [`registry::known_kinds`](./registry.rs), and because `store` mints an agent's handle (at launch and in the snapshot fold) and validates every candidate name through `agents::petname`.
 - Shared decisions keep one home, and adapters supply the inputs: [`hook_types.rs`](./hook_types.rs)'s `SessionSource::session_start_signal` owns what a `SessionStart` source means (`compact` closes a compaction, every other source registers), and `LocalContextRefreshCtx::changed_transcript` owns the unchanged-source gate. An adapter parses the native source and resolves the logical stat that stands for its local source; neither rule is restated per kind.
 - Private provider implementations live below [`adapters/`](./adapters/mod.rs). Each built-in directory owns its native payloads, parsers, installer, probes, and selected capability implementations; [`adapters/plugin/`](./adapters/plugin/mod.rs) loads validated machine-tier process plugins, while [`plugins.rs`](./plugins.rs) is the public provider-neutral plugin façade.
 - `spend.rs` is the read-only, sidebar-safe full-history cost parser; a CI grep keeps every `spend.rs` free of store-write, run-wake, and broker imports.
