+++ crates/rimz/src/cli/hooks/lifecycle/identity.rs
@@ -84,7 +84,7 @@ pub(in crate::cli::hooks) fn fill_root_launch_identity(
 }
 
 pub(super) fn validate_agent_name_env(raw: String, source: &str, _var: &str) -> Option<String> {
-    if rimz::harness::petname::valid_agent_name(&raw) {
+    if rimz::agents::petname::valid_agent_name(&raw) {
         Some(raw)
     } else {
         warn!(
