+++ crates/rimz/src/harness/AGENTS.md
@@ -6,12 +6,12 @@ Topic detail lives in [fleet.md](../../../../docs/internals/harness/fleet.md) (s
 
 ## Boundaries
 
-- The harness owns layout specs, effective launch resolution, profile prompt-file validation, launch finalization, provider-process compilation and argv, launch reminder assembly (`launch_reminders.rs` owns the one `<system_reminder>` tag, its paragraph order, and the model line; the team and catalog producers return bodies and never wrap), stable handles, supervised run records, the durable blocking-wait policy and run-wake socket (`run_wake.rs`), rebirth inspection and materialization, cohort and lane qualification/resume planning, relaunch posture resolution (`resume::resolve_posture`, shared by restart, fork, and every resume path), lane restore materialization, loop scheduling, hidden runner domain, auto-continue policy, and idle-compaction policy.
+- The harness owns layout specs, effective launch resolution, profile prompt-file validation, launch finalization, provider-process compilation and argv, launch reminder assembly (`launch_reminders.rs` owns the one `<system_reminder>` tag, its paragraph order, and the model line; the team and catalog producers return bodies and never wrap), supervised run records, the durable blocking-wait policy and run-wake socket (`run_wake.rs`), rebirth inspection and materialization, cohort and lane qualification/resume planning, relaunch posture resolution (`resume::resolve_posture`, shared by restart, fork, and every resume path), lane restore materialization, loop scheduling, hidden runner domain, auto-continue policy, and idle-compaction policy.
 - Harness delivery rides `message`; it never reimplements queues, gates, retries, or transcript audit.
 - `schedule/signal.rs` owns the signal vocabulary: the `SignalName` grammar and `SignalSelector`, the `agent`, `wake`, `team`, `ci`, and `pr` families reserved from user emitters, the agent and team lifecycle mappings, and `fire_signal`. A subscription resolves against the signal's whole family: an exact name delivers, another member of the family skips (recorded, no delivery), and a different family or a failed `--match` is ignored. Firing is in-process, with no queue and no replay: whichever process observes the event fires the tasks armed for it right there, and a signal nobody subscribes to is gone. A detached watcher is single-flight through its `RuntimePaths` lock, and losing that lock is an outcome the elder resolves, never something the watcher retries. `run_watcher` is that whole watcher lifetime in one function (catalog gate, lock, check, signal append, fire), and it holds the lock through the fire so the elder cannot call the row lost mid-flight; the CLI only spawns the detached child and hands over the store and workspace.
 - `schedule/runner.rs` owns the stop ladder as `stop_task`: run-lock probe, durable cancellation, grace wait, SIGTERM, grace wait, the `canceled` history row, and the manual-retry error. The CLI supplies the cancellation as a closure and prints the outcome.
 - Terminal loop history rows are appended in the harness only, through the one `run_log` transition; the CLI renders the transition it gets back and never writes a row.
-- Handle rendering and address resolution read the addressable peer set; raw root rows are for accounting scans.
+- Handle rendering and address resolution read the addressable peer set; raw root rows are for accounting scans. Minting a handle and deciding whether a name is legal belong to `agents::petname`, which the harness calls.
 - CLI handlers keep argument parsing, presentation, and cross-command orchestration. Harness modules own pure domain rules, durable records, helper argv shape, and side-effect boundaries.
 - Elder-fired helpers in `schedule/fire.rs`, `auto_continue.rs`, `auto_redeem.rs`, and `idle_compact.rs` spawn hidden CLI subprocesses with fresh null stdio.
 - `auto_continue.rs`, `auto_redeem.rs`, `budget.rs`, `idle_compact.rs`, `orphan_sweep.rs`, and `run_timeout.rs` are in the sidebar import graph. Keep their evaluation side read-only and delegate durable mutations or provider transport to detached helpers; runtime-cache writes through `disk::atomic::write_temp_then_rename_cache` are the allowed durability path.
